Implement MyTokenizer function
strategy is to use the default tokenizer in gTTS to do the heavy lifting
but check which sentences (or phrases) exceed the 100-character limit of
the Google API. For those long sentences (or phrases), chunking is
accomplished with the nltk library RegexpParser. Chunks considered are
variations of noun phrases, noun+prepositional phrases, compound nouns,
and verb phrases. Shorter phrases are reconstructed from these chunks,
with the assumption that no resulting chunk will exceed 100 characters.